requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.19.2
pyyaml>=5.4.1
scikit-learn>=0.24.2
pillow>=8.3.1
opencv-python>=4.5.3
matplotlib>=3.4.3
